myenv/RentCartApi/requirements.txt
django-cors-headers=3.14.0
djangorestframework-simplejwt=5.2.2       
Django=4.2.1    
djangorestframework=3.14.0   